ADD pc-03/hw01

+++ pc-03/hw01/point.h
@@ -8,11 +8,11 @@ class POINT
         void Output();
         POINT();
         POINT(const POINT &);
-        POINT (int, int);
+        POINT (double, double);
         ~POINT();
-        void SetX(const int &);
-        void SetY(const int &);
-        int GetX();
-        int GetY();
-        void Move(const int &, const int &);
+        void SetX(const double &);
+        void SetY(const double &);
+        double GetX();
+        double GetY();
+        void Move(const double &, const double &);
 };
